requirements.txt
pyyaml
mechanize
selenium
